Migrate queries to gamedata
