+++ eos/db/saveddata/__init__.py
@@ -1,19 +1,5 @@
 __all__ = [
-    "character",
-    "fit",
-    "module",
-    "user",
-    "skill",
-    "price",
-    "booster",
-    "drone",
-    "implant",
-    "fleet",
-    "damagePattern",
-    "miscData",
-    "targetResists",
-    "override",
-    "crest",
-    "implantSet",
-    "loadDefaultDatabaseValues"
+    "loadDefaultDatabaseValues",
+    "mapper",
+    "queries",
 ]
